auth_microservice/pyproject.toml
[tool.poetry]
name = "auth-microservice"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.2"}
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.29.0"
bcrypt = "^4.2.0"
pydantic-settings = "^2.6.0"
pytest = "^8.3.3"
alembic = "^1.13.3"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
